Hide test_kms_azure_online behind ENABLE_ONLINE_TESTS

+++ kms-message/test/test_kms_azure_online.c
@@ -161,7 +161,7 @@ send_kms_request (kms_request_t *req, const char *host)
    while ((bytes_to_read =
               kms_response_parser_wants_bytes (response_parser, 1024)) > 0) {
       bytes_read =
-         mongoc_stream_read (tls_stream, buf, 1024, 0, socket_timeout_ms);
+         (int) mongoc_stream_read (tls_stream, buf, 1024, 0, socket_timeout_ms);
       if (!kms_response_parser_feed (response_parser, buf, bytes_read)) {
          TEST_ERROR ("read failed: %s",
                      kms_response_parser_error (response_parser));
